chore: Move ILogger interface into a separate file
